feat!: Add `gix-shallow` crate and use it from `gix` and `gix-protocol`
That way it's easier to reuse shallow-handling code from plumbing crates.

Note that this is a breaking change as `gix-protocol` now uses the `gix-shallow::Update`
type, which doesn't implement a formerly public `from_line()` method anymore.
Now it is available as `fetch::response::shallow_update_from_line()`.
