+++ Cargo.toml
@@ -304,6 +304,7 @@ members = [
     "gix-ref/tests",
     "gix-config/tests",
     "gix-traverse/tests",
+    "gix-shallow"
 ]
 
 [workspace.dependencies]
