+++ gix/Cargo.toml
@@ -328,6 +328,7 @@ gix-dir = { version = "^0.10.0", path = "../gix-dir", optional = true }
 gix-config = { version = "^0.42.0", path = "../gix-config" }
 gix-odb = { version = "^0.65.0", path = "../gix-odb" }
 gix-hash = { version = "^0.15.1", path = "../gix-hash" }
+gix-shallow = { version = "^0.1.0", path = "../gix-shallow" }
 gix-object = { version = "^0.46.0", path = "../gix-object" }
 gix-actor = { version = "^0.33.1", path = "../gix-actor" }
 gix-pack = { version = "^0.55.0", path = "../gix-pack", default-features = false, features = [
